API

SimpleNNs.Conv — Method

Conv(kernel_size::NTuple{N, Int}, out_channels::Int; kwargs...)

A convolutional layer with a given kernel size and specified number of output channels.

This can automatically infer the number of input channels based on the preceeding layers.

Keyword Arguments

use_bias (default: Val(true)) - Whether or not to add a bias vector to the output. Wrapped in a Val for optimisation.

activation_fn (default: identity) - A custom activation function. Note that not all functions are supported by backpropagation.

parameter_type (default: Val(Float32)) - The datatype to use for the parameters, wrapped in a Val type.

SimpleNNs.Dense — Method

Dense(outputs::Integer; kwargs...)

A representation of a dense layer. By default this can be constructed by specifying the desired number of outputs. The input size can be inferred from the rest of the chain when constructing a model.

Keyword Arguments

use_bias (default: Val(true)) - Whether or not to add a bias vector to the output. Wrapped in a Val for optimisation.

activation_fn (default: identity) - A custom activation function. Note that not all functions are supported by backpropagation.

parameter_type (default: Val(Float32)) - The datatype to use for the parameters, wrapped in a Val type.

inputs (default: Infer()) - Specify the number of inputs, or infer them from the rest of the model.

SimpleNNs.Flatten — Type

Flatten()

Flatten the dimensions of the preceeding layer, leaving the batch dimension unaffected. The output should be (k x n) where k is the product of the non-batch dimensions of the previous layer.

SimpleNNs.LogitCrossEntropyLoss — Type

LogitCrossEntropyLoss(targets, num_classes::Int)

Expects the targets in a single vector containg class labels, which have to be between 1 and num_classes inclusive.

SimpleNNs.MSELoss — Type

MSELoss(targets)

Expects the targets in the form (K x N) where K is the output dimension (usually 1) and N is the batch size.

SimpleNNs.MaxPool — Method

MaxPool(pool_size::NTuple{N, Int}; kwargs...)

A convolutional max-pool layer with a given kernel size.

This can automatically infer the necessary sizes if specified.

SimpleNNs.Static — Method

Static(inputs::Union{Int, NTuple}; kwargs...)

Used for specifying the input type to a neural network. inputs should be a single integer for a dense network, representing the number of features. For a image network, inputs can be a tuple specifying the size of the images in the form (WIDTH, HEIGHT, CHANNELS).

SimpleNNs.activation_gradient_fn — Method

Dertivatives are used to backpropagate the gradients of the layer outputs back to the activations of that layer. To save space, these are calculated exclusively using the outputs of the layer. Instead of functions written as dy/dx=f(x), we instead write dy/dx = g(y). This can be done for the 3 major functions.

Whenever y is used below, assume this is a function of the output, not the input.

SimpleNNs.backprop! — Method

backprop!(partials_buffer, gradient_buffer, inputs, outputs, layer)

Backpropagates the partial gradients of the outputs of the current layer into the parameters of the current layer. partial_buffers is used as a buffer for the gradients of the output of this layer. gradient_buffer should be filled up with the gradients of the parameters of the current layer, using the chain rule. inputs is the array fed into the layer and outputs is the output of this layer in the forward pass. layer is the struct containing information about the layer.

SimpleNNs.chain — Method

chain(layers...)

Combines the given layer definitions into a single model and propagates the layer sizes through the network.

The first layer must always be a Static layer which specifies the feature size. If this is a simple fully connection network, then the first layer should be Static(nf) where nf is the number of features in your input matrix. Do not specify the batch size in this static input.

The default datatype for most layers is Float32, but this may be changed. The parameters of the entire model must be of the same datatype. This function will create a flat parameter vector for the model which can be accessed using the parameters function.

Examples

A simple dense, fully-connected, neural network which has 3 input features:

model = chain(
    Static(3),
    Dense(10, activation_fn=tanh),
    Dense(10, activation_fn=sigmoid),
    Dense(1, activation_fn=identity),
);

An example convolutional neural network:

# Image size is (WIDTH, HEIGHT, CHANNELS)
img_size = (28, 28, 1)
model = chain(
    Static(img_size),
    Conv((5,5), 16; activation_fn=relu),
    MaxPool((2,2)),
    Conv((3,3), 8; activation_fn=relu),
    MaxPool((4,4)),
    Flatten(),
    Dense(10, activation_fn=identity)
)

See also Static, Dense, Conv, MaxPool, Flatten and preallocate.

SimpleNNs.get_outputs — Method

get_outputs(cache::ForwardPassCache)

Gets the last output from the forward pass buffer.

SimpleNNs.gradients — Method

gradients(cache::BackpropagationCache)

Extracts the gradient array from the backwards pass buffer, filled from use of the backprop! function.

SimpleNNs.parameters — Method

parameters(model::Model)

Returns the array used to store the parameters of the model.

Modifying this array will change the parameters of the model.

SimpleNNs.preallocate — Method

preallocate(model::Model, batch_size::Integer)

Creates a buffer to store the intermediate layer outputs of a forward pass, along with the input.

The inputs can be set using set_inputs! and the outputs can be retrieved using get_outputs.

SimpleNNs.preallocate_grads — Method

preallocategrads(model::Model, batchsize::Integer)

Creates a buffer to store the intermediate arrays needed for backpropagation.

The gradients can be retrieved from the buffer using gradients on the buffer.

SimpleNNs.pullback! — Method

pullback!(input_partials, output_partials, layer)

Here, we complete the backpropagation of the partial gradients to the inputs of the current layer. This should be called after backprop!. This method will fill the input_partials buffer with partial gradients calculated via the chain rule from the gradients of the partials from this layer's output.

SimpleNNs.set_inputs! — Method

set_inputs!(cache::ForwardPassCache, inputs)

Sets the input array in the forward pass cache.

SimpleNNs.GPU.gpu — Method

gpu(x)

Creates a copy of x on the GPU, using CUDA. Works for models created with chain or plain arrays.